@@ -21,14 +21,15 @@
         }
     },
     "require": {
-        "php": ">=7.4",
-        "symfony/config": "^4.4 || ^5.0 || ^6.0",
-        "symfony/dependency-injection": "^4.4 || ^5.0 || ^6.0",
-        "symfony/http-kernel": "^4.4 || ^5.0 || ^6.0",
-        "symfony/yaml": "^4.4 || ^5.0 || ^6.0",
-        "spiral/roadrunner": "^2.0.2",
-        "spiral/roadrunner-worker": "^2.1.4",
-        "spiral/goridge": "^3.0",
+        "php": ">=8.1",
+        "symfony/config": "^5.4 || ^6.0",
+        "symfony/dependency-injection": "^5.4 || ^6.0",
+        "symfony/http-kernel": "^5.4 || ^6.0",
+        "symfony/yaml": "^5.4 || ^6.0",
+        "spiral/roadrunner": "^2023.3.0",
+        "spiral/roadrunner-http": "^3.0",
+        "spiral/roadrunner-worker": "^3.0",
+        "spiral/goridge": "^4.0",
         "symfony/psr-http-message-bridge": "^1.1 || ^2.0",
         "psr/log": "^1.1 || ^2.0 || ^3.0"
     },
@@ -37,7 +38,8 @@
         "symfony/proxy-manager-bridge": "For doctrine re-connection implementation"
     },
     "require-dev": {
-        "symfony/var-dumper": "^4.0 || ^5.0 || ^6.0",
+        "symfony/var-dumper": "^5.4 || ^6.0",
+        "symfony/var-exporter": "^5.4 || ^6.0",
         "phpunit/phpunit": "^9.5",
         "phpspec/prophecy": "^1.11",
         "phpspec/prophecy-phpunit": "^2.0",
@@ -50,12 +52,12 @@
         "blackfire/php-sdk": "^1.21",
         "doctrine/doctrine-bundle": "^2.1.1",
         "doctrine/orm": "^2.7.3",
-        "symfony/proxy-manager-bridge": "^4.0 || ^5.0 || ^6.0",
-        "spiral/roadrunner-metrics": "^2.0.1",
+        "symfony/proxy-manager-bridge": "^5.4 || ^6.0",
+        "spiral/roadrunner-metrics": "^3.1",
         "mikey179/vfsstream": "^1.6.8",
-        "symfony/runtime": "^5.3.0 || ^6.0",
+        "symfony/runtime": "^5.4 || ^6.0",
         "laminas/laminas-zendframework-bridge": "^1.4",
-        "spiral/roadrunner-grpc": "^2.0"
+        "spiral/roadrunner-grpc": "^3.0"
     },
     "conflict": {
         "doctrine/doctrine-bundle": "<2.1.1",
