@@ -17,7 +17,6 @@ use Baldinof\RoadRunnerBundle\Reboot\KernelRebootStrategyInterface;
 use Baldinof\RoadRunnerBundle\RoadRunnerBridge\HttpFoundationWorker;
 use Baldinof\RoadRunnerBundle\RoadRunnerBridge\HttpFoundationWorkerInterface;
 use Baldinof\RoadRunnerBundle\Worker\GrpcWorker as InternalGrpcWorker;
-use Baldinof\RoadRunnerBundle\Worker\GrpcWorkerInterface;
 use Baldinof\RoadRunnerBundle\Worker\HttpDependencies;
 use Baldinof\RoadRunnerBundle\Worker\HttpWorker as InternalHttpWorker;
 use Baldinof\RoadRunnerBundle\Worker\WorkerRegistry;
@@ -37,14 +36,6 @@ use Spiral\RoadRunner\Worker as RoadRunnerWorker;
 use Spiral\RoadRunner\WorkerInterface as RoadRunnerWorkerInterface;
 use Symfony\Component\HttpKernel\Kernel;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
-
-// Polyfill of the `service()` function introduced in Symfony 5.1 when using older version
-if (!\function_exists('Symfony\Component\DependencyInjection\Loader\Configurator\service')) {
-    function service(string $id): ReferenceConfigurator
-    {
-        return ref($id); // @phpstan-ignore-line
-    }
-}
 
 return static function (ContainerConfigurator $container) {
     $container->parameters()
@@ -136,7 +127,7 @@ return static function (ContainerConfigurator $container) {
                 service(GrpcInvoker::class),
             ]);
 
-        $services->set(GrpcWorkerInterface::class, InternalGrpcWorker::class)
+        $services->set(InternalGrpcWorker::class)
             ->public() // Manually retrieved on the DIC in the Worker if the kernel has been rebooted
             ->tag('monolog.logger', ['channel' => BaldinofRoadRunnerExtension::MONOLOG_CHANNEL])
             ->args([
@@ -150,7 +141,7 @@ return static function (ContainerConfigurator $container) {
             ->get(WorkerRegistryInterface::class)
             ->call('registerWorker', [
                 Environment\Mode::MODE_GRPC,
-                service(GrpcWorkerInterface::class),
+                service(InternalGrpcWorker::class),
             ]);
     }
 };
