@@ -77,7 +77,10 @@ class Configuration implements ConfigurationInterface
                                     ->scalarNode('subsystem')->defaultNull()->end()
                                     ->arrayNode('labels')
                                         ->scalarPrototype()
-                                            ->validate()->always(fn ($value) => (string) $value)->end()
+                                            ->validate()
+                                                ->ifEmpty()->thenInvalid('Metric label value cannot be empty')
+                                                ->always(fn ($value) => (string) $value)
+                                            ->end()
                                         ->end()
                                     ->end()
                                     ->arrayNode('buckets')
